+++ Comex/Comex.Modelos/Produtos/ProdutoIsento.cs
@@ -6,14 +6,14 @@ namespace Comex.Modelos.Produtos
     public class ProdutoIsento : Produto
     {
         public static new int Quantidade { get; set; }
-        public ProdutoIsento(string nome, string descricao, double preco, int estoque, Categoria categoria)
+        public ProdutoIsento(string nome, string descricao, decimal preco, int estoque, Categoria categoria)
             : base(nome, descricao, preco, estoque, categoria)
         {
             Quantidade++;
             Id = Quantidade;
         }
 
-        public override double CalculaImposto()
+        public override decimal CalculaImposto()
         {
             return 0;
         }
